refactor(racingCar): 비지니스 로직과 UI 로직 분리
