--- src/main/java/racingcar/Application.java
@@ -1,10 +1,0 @@
-package racingcar;
-
-public class Application {
-    public static void main(String[] args) {
-        Car car = new Car();
-        car.displayCarInformation();
-        car.displayWhoIsWinner();
-    }
-}
-
