+++ src/main/java/racingcar/domain/Application.java
@@ -1,0 +1,14 @@
+package racingcar.domain;
+
+public class Application {
+    public static void main(String[] args) {
+        User user = new User();
+        int moveCount = user.howMuchTry();
+
+        while (moveCount>=0){
+            moveCount--;
+        }
+
+    }
+}
+
