+++ src/main/java/racingcar/domain/Car.java
@@ -1,0 +1,11 @@
+package racingcar.domain;
+
+import java.util.*;
+
+import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
+import static camp.nextstep.edu.missionutils.Console.readLine;
+
+public class Car{
+
+}
+
