+++ src/main/java/racingcar/view/OutputView.java
@@ -1,0 +1,7 @@
+package racingcar.view;
+
+public class OutputView {
+    public void result(){
+        System.out.println("실행 결과");
+    }
+}
